This commit mostly fixes unit tests
Unit tests have mostly been fixed to work with the transition to BTCZ.
A few of the tests have been disabled because I do not fully understand
the scope of the tests. If someone has a better understanding they can
take a look at them and see what they can do to get them to a healthy
state.